move code from data_src_dependent to DataFormatter, to match convention
