+++ src/main/java/bridges/data_src_dependent/BridgesAssignment.java
@@ -1,56 +1,20 @@
 package bridges.data_src_dependent;
+
+import bridges.base.DataStruct;
+import bridges.connect.Bridges;
+
+import java.util.List;
+
 /**
- * @brief  A class which fetches assignments previously posted to BRIDGES server and reconstructs the corresponding
- * BRIDGES object / data structure.
+ * TODO Think about if it is worth it to capture an entire assignment, storing subassignments in a List of some sort
  *
  * @author Alec Goncharow
  * @date   12/29/18
  *
  */
 public class BridgesAssignment extends DataSource {
-
-	/**Reconstruct a ColorGrid from an existing ColorGrid on the Bridges server
-	 *
-	 * @return the ColorGrid stored in the bridges server
-	 * @param user the name of the user who uploaded the assignment
-	 * @param assignment the ID of the assignment to get
-	 * @param subassignment the ID of the subassignment to get
-	 **/
-	static bridges.base.ColorGrid getColorGridFromAssignment(String user, int assignment, int subassignment) {
-
-		return null;
-    }
-
-    /**Reconstruct a ColorGrid from an existing ColorGrid on the Bridges server
-	 *
-	 * @return the ColorGrid stored in the bridges server
-	 * @param user the name of the user who uploaded the assignment
-	 * @param assignment the ID of the assignment to get
-	 **/
-    static bridges.base.ColorGrid getColorGridFromAssignment(String user, int assignment) {
-		return getColorGridFromAssignment(user, assignment, 0);
-	}
-
-	/***
-	 * This function obtains the JSON representation of a particular subassignment.
-	 *
-	 * @return a string that is the JSON representation of the subassignment as stored by the Bridges server.
-	 * @param user the name of the user who uploaded the assignment
-	 * @param assignment the ID of the assignment to get
-	 * @param subassignment the ID of the subassignment to get
-	 ***/
-	private static String getAssignment(String user, int assignment, int subassignment) {
-
-		return null;
-	}
-	/***
-	 * This function obtains the JSON representation of the first subassignment of an assignment.
-	 *
-	 * @return a string that is the JSON representation of the subassignment as stored by the Bridges server.
-	 * @param user the name of the user who uploaded the assignment
-	 * @param assignment the ID of the assignment to get
-	 ***/
-	private static String getAssignment(String user, int assignment) {
-        return getAssignment(user, assignment, 0);
-	}
+    private String username, assignmentName, assignmentNumber, title, description, vistype;
+    private int assignmentID;
+    private List<DataStruct> assignments;
+    public BridgesAssignment() {}
 }
